(Partially) fixed the error with the sell button
The sell button is now responsive, although it is still error prone.
Considering I'm not sure exactly what direction you wanted to go with
the button I left that up for you to decide.

I would point out though, you have an undeclared variable on line 156 of
the the stock-index.js file. You declare the variable "sellTotal"; is
that supposed to be distinct from "sellValue", or is it a variable name
that didn't get changed in a re-factor?

I will create an issue for you to document the undefined behavior of the
sell button in more detail.